Added Assessment editors, Assessment Locks, and assessment editor/owner check.

+++ assessments_display_all.php
@@ -27,7 +27,7 @@
 	if($pagerestrictions=="")
 	{
 		
-		$sql = "SELECT * FROM assessments where Owner='".$_SESSION['useremail']."' order by Title";
+		$sql = "SELECT * FROM assessments where Owner='".$_SESSION['useremail']."' or Editors LIKE '%".$_SESSION['useremail']."%' order by Title";
 		$result = $db->query($sql);
 		$rowcount=mysqli_num_rows($result);
 		if($rowcount!=0)
@@ -50,7 +50,7 @@
 							<tbody>
 								
 								<?php
-								$sql = "SELECT * FROM assessments where Owner='".$_SESSION['useremail']."' order by Title";
+								$sql = "SELECT * FROM assessments where Owner='".$_SESSION['useremail']."' or Editors LIKE '%".$_SESSION['useremail']."%' order by Title";
 								$result = $db->query($sql);
 								while($row = $result->fetch_assoc())
 								{
@@ -59,6 +59,9 @@
 									$Description=htmlspecialchars($row["Description"], ENT_QUOTES);
 									$Subject=htmlspecialchars($row["Subject"], ENT_QUOTES);
 									$Grade=htmlspecialchars($row["Grade"], ENT_QUOTES);
+									$Locked=htmlspecialchars($row["Locked"], ENT_QUOTES);
+									$Owner=htmlspecialchars($row["Owner"], ENT_QUOTES);
+									$Editors=htmlspecialchars($row["Editors"], ENT_QUOTES);
 									$Code=htmlspecialchars($row["Code"], ENT_QUOTES);
 								
 									echo "<tr class='assessmentrow'>";
@@ -71,9 +74,12 @@
 											echo "<div class='morebutton' style='position:absolute; margin-top:-15px;'>";
 												echo "<button id='demo-menu-bottom-left-$Assessment_ID' class='mdl-button mdl-js-button mdl-button--icon mdl-js-ripple-effect mdl-color-text--grey-600'><i class='material-icons'>more_vert</i></button>";
 												echo "<ul class='mdl-menu mdl-menu--bottom-right mdl-js-menu mdl-js-ripple-effect' for='demo-menu-bottom-left-$Assessment_ID'>";
-													echo "<li class='mdl-menu__item modal-createassessment' href='#createassessment' data-title='$Title' data-description='$Description' data-subject='$Subject' data-assessmentid='$Assessment_ID' data-grade='$Grade' style='font-weight:400'>Edit Assessment</a></li>";
-													echo "<li class='mdl-menu__item exploreassessment'><a href='#assessments/$Assessment_ID' class='mdl-color-text--black' style='font-weight:400'>Edit Questions</a></li>";
-													echo "<li class='mdl-menu__item deleteassessment'><a href='modules/".basename(__DIR__)."/assessment_delete.php?assessmentid=".$Assessment_ID."' class='mdl-color-text--black' style='font-weight:400'>Delete</a></li>";
+													echo "<li class='mdl-menu__item exploreassessment'><a href='#assessments/$Assessment_ID' class='mdl-color-text--black' style='font-weight:400'>Questions</a></li>";
+													if($Owner==$_SESSION['useremail'])
+													{
+														echo "<li class='mdl-menu__item modal-createassessment' href='#createassessment' data-title='$Title' data-description='$Description' data-subject='$Subject' data-assessmentid='$Assessment_ID' data-grade='$Grade' data-editors='$Editors' data-locked='$Locked' style='font-weight:400'>Settings</a></li>";
+														echo "<li class='mdl-menu__item deleteassessment'><a href='modules/".basename(__DIR__)."/assessment_delete.php?assessmentid=".$Assessment_ID."' class='mdl-color-text--black' style='font-weight:400'>Delete</a></li>";
+													}
 												echo "</ul>";
 											echo "</div>";
 	
@@ -118,7 +124,18 @@
 			$(".modal-content #assessment_title").val(Assessment_Title);
 			var Assessment_Description = $(this).data('description');
 			$(".modal-content #assessment_description").val(Assessment_Description);
+			var Assessment_Editors = $(this).data('editors');
+			$(".modal-content #assessment_editors").val(Assessment_Editors);
 			var Assessment_Grade = $(this).data('grade');
+			var Assessment_Locked = $(this).data('locked');
+			if(Assessment_Locked=='1')
+			{
+				$(".modal-content #assessment_lock").prop('checked',true);
+			}
+			else
+			{
+				$(".modal-content #assessment_lock").prop('checked',false);
+			}
 			if(Assessment_Grade!="blank")
 			{
 				var Assessment_Grade_String=String(Assessment_Grade);
@@ -150,7 +167,7 @@
 		//Delete assessment
 		$( ".deleteassessment" ).unbind().click(function() {
 			event.preventDefault();
-			var result = confirm("Delete this assessment?");
+			var result = confirm("Are you sure you want to delete this assessment?");
 			if (result) {
 
 				//Make the post request
